+++ sdk/appcenter-distribute/src/main/java/com/microsoft/appcenter/distribute/download/manager/DownloadManagerRequestTask.java
@@ -45,11 +45,19 @@ class DownloadManagerRequestTask extends AsyncTask<Void, Void, Void> {
             request.setVisibleInDownloadsUi(false);
         }
         long enqueueTime = System.currentTimeMillis();
-        long downloadId = downloadManager.enqueue(request);
-        if (isCancelled()) {
-            return null;
+        try {
+            long downloadId = downloadManager.enqueue(request);
+            if (!isCancelled()) {
+                mDownloader.onDownloadStarted(downloadId, enqueueTime);
+            }
+        } catch (IllegalArgumentException e) {
+            
+            /*
+             * In cases when Download Manager application is disabled,
+             * IllegalArgumentException: Unknown URL content://downloads/my_download is thrown.
+             */
+            mDownloader.onDownloadError(new IllegalStateException("Failed to start download: Download Manager is disabled.", e));
         }
-        mDownloader.onDownloadStarted(downloadId, enqueueTime);
         return null;
     }